Title: 바구니 안의 사탕, Time: 1294ms, Memory: 57MB, Status: Wrong Answer - CodeTree

+++ 250224/바구니 안의 사탕/candy-in-the-basket.java
@@ -23,8 +23,8 @@ public class Main {
         }
 
         int right = min;
-        int cnt = 0;
-        int ans = 0;
+        long cnt = 0;
+        long ans = 0;
         for(int left=min; left<=max-2*k; left++){
             while(right <= left+2*k){
                 if(map.containsKey(right)){
